define rc and bfields in a separate source, keep same header
